Bundle jointjs in pipeline-builder
These changes reconfigure rollup so it bundles jointjs into out package.

The reason for this bundnling is the fact that jointjs sticks to
lodash.3.10.1 while we are using the latest version due to new features
and bugfixes. Therefore, the version conflict may appear when using our
library. Thus, we decided to embed jointjs.